Remove unused tilelang.common package
