--- tilelang/common/__init__.py
@@ -1,1 +1,0 @@
-from .transform_kind import TransformKind  # noqa: F401
