+++ tilelang/cuda/intrinsics/macro/mma_macro_generator.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
+from enum import IntEnum
 import tilelang.language as T
 from typing import Literal
 from collections.abc import Callable
-from tilelang.common import TransformKind
 from tvm import DataType
 from tvm import tirx
 from tvm.ir import Range
@@ -31,6 +31,14 @@ from tilelang.cuda.intrinsics.layout.mma_layout import (
 )
 
 lift = convert
+
+
+# Copied from bitblas
+class TransformKind(IntEnum):
+    NonTransform = 0
+    InterWarpTransform = 1
+    IntraWarpTransform = 2
+    LDMatrixTransform = 3
 
 
 class TensorCoreIntrinEmitter:
